chore(earn): Clean up single pool earn flow (#6133)
### Description

Remove all references of SHOW_MULTIPLE_EARN_POOLS and
SHOW_STABLECOIN_EARN. Remove pool card from the Aave pool token details
screen.

Follow up to remove Aave pool token from token details list:
https://linear.app/valora/issue/ACT-1404/dont-show-aave-pool-token-on-token-list-version-=195

### Test plan

Unit tests updated, CI and knip.

Ran wallet locally and went through multi-pool earn screens, nothing
broken, no warnings or errors.

### Related issues

- Fixes ACT-1365

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)
