+++ src/analytics/Events.tsx
@@ -653,7 +653,6 @@ export enum PointsEvents {
 }
 
 export enum EarnEvents {
-  earn_cta_press = 'earn_cta_press',
   earn_entrypoint_press = 'earn_entrypoint_press',
   earn_before_deposit_action_press = 'earn_before_deposit_action_press',
   earn_deposit_provider_info_press = 'earn_deposit_provider_info_press',
@@ -664,12 +663,7 @@ export enum EarnEvents {
   earn_deposit_submit_success = 'earn_deposit_submit_success',
   earn_deposit_submit_error = 'earn_deposit_submit_error',
   earn_deposit_submit_cancel = 'earn_deposit_submit_cancel',
-  earn_view_pools_press = 'earn_view_pools_press',
-  earn_enter_amount_info_press = 'earn_enter_amount_info_press',
   earn_enter_amount_continue_press = 'earn_enter_amount_continue_press',
-  earn_enter_amount_info_more_pools = 'earn_enter_amount_info_more_pools',
-  earn_exit_pool_press = 'earn_exit_pool_press',
-  earn_deposit_more_press = 'earn_deposit_more_press',
   earn_deposit_add_gas_press = 'earn_deposit_add_gas_press',
   earn_feed_item_select = 'earn_feed_item_select',
   earn_collect_earnings_press = 'earn_collect_earnings_press',
