--- src/earn/constants.ts
@@ -1,1 +1,0 @@
-export const PROVIDER_ID = 'aave-v3' as const
