+++ src/earn/hooks.ts
@@ -14,7 +14,6 @@ import { useCashInTokens, useSwappableTokens } from 'src/tokens/hooks'
 import { TokenBalances } from 'src/tokens/slice'
 import Logger from 'src/utils/Logger'
 import { ensureError } from 'src/utils/ensureError'
-import networkConfig from 'src/web3/networkConfig'
 
 const TAG = 'earn/hooks'
 
@@ -34,19 +33,6 @@ export function useEarnPositionProviderName(providerId: string) {
     Logger.warn(TAG, 'providerName not found', providerId)
   }
   return providerName
-}
-
-// Helper hook to get position given a positionId. Defaults to the aave position
-// while we're in the interim period of building the multiple pool flow
-export function useEarnPosition(positionId: string = networkConfig.aaveArbUsdcTokenId) {
-  const pools = useSelector(earnPositionsSelector)
-  return useMemo(() => {
-    const pool = pools.find((pool) => pool.positionId === positionId)
-    if (!pool) {
-      Logger.warn(TAG, 'pool not found', positionId)
-    }
-    return pool
-  }, [pools, positionId])
 }
 
 export function useDepositEntrypointInfo({
