+++ src/tokens/TokenDetails.test.tsx
@@ -7,15 +7,12 @@ import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { Price } from 'src/priceHistory/slice'
 import { getFeatureGate } from 'src/statsig'
-import { StatsigFeatureGates } from 'src/statsig/types'
 import TokenDetailsScreen from 'src/tokens/TokenDetails'
 import { NetworkId } from 'src/transactions/types'
 import { ONE_DAY_IN_MILLIS } from 'src/utils/time'
-import networkConfig from 'src/web3/networkConfig'
 import MockedNavigator from 'test/MockedNavigator'
 import { createMockStore } from 'test/utils'
 import {
-  mockAaveArbUsdcAddress,
   mockCeloTokenId,
   mockPoofTokenId,
   mockTestTokenTokenId,
@@ -262,38 +259,6 @@ describe('TokenDetails', () => {
     )
 
     expect(queryByText('celoNews.headerTitle')).toBeTruthy()
-  })
-
-  it('renders earn card for Aave Arbitrum Usdc when user has balance', () => {
-    jest
-      .mocked(getFeatureGate)
-      .mockImplementation((featureGate) => featureGate === StatsigFeatureGates.SHOW_STABLECOIN_EARN)
-    const store = createMockStore({
-      tokens: {
-        tokenBalances: {
-          [networkConfig.aaveArbUsdcTokenId]: {
-            networkId: NetworkId['arbitrum-sepolia'],
-            address: mockAaveArbUsdcAddress,
-            tokenId: networkConfig.aaveArbUsdcTokenId,
-            symbol: 'aArbSepUSDC',
-            priceUsd: '1',
-            balance: '10.74',
-            priceFetchedAt: Date.now(),
-          },
-        },
-      },
-    })
-
-    const { getByTestId } = render(
-      <Provider store={store}>
-        <MockedNavigator
-          component={TokenDetailsScreen}
-          params={{ tokenId: networkConfig.aaveArbUsdcTokenId }}
-        />
-      </Provider>
-    )
-
-    expect(getByTestId('EarnActivePool')).toBeTruthy()
   })
 
   it('does not render chart if no prices are found and error status', () => {
