+++ src/web3/networkConfig.ts
@@ -96,9 +96,6 @@ interface NetworkConfig {
   spendTokenIds: string[]
   saveContactsUrl: string
   getPointsConfigUrl: string
-  arbAavePoolV3ContractAddress: Address
-  arbAaveIncentivesV3ContractAddress: Address
-  aaveArbUsdcTokenId: string
   internalRpcUrl: Record<Network.Arbitrum, string>
   authHeaderIssuer: string
   web3AuthVerifier: string
@@ -157,16 +154,6 @@ const ETH_TOKEN_ID_MAINNET = `${NetworkId['ethereum-mainnet']}:native`
 
 const ARB_USDC_TOKEN_ID_STAGING = `${NetworkId['arbitrum-sepolia']}:0x75faf114eafb1bdbe2f0316df893fd58ce46aa4d`
 const ARB_USDC_TOKEN_ID_MAINNET = `${NetworkId['arbitrum-one']}:0xaf88d065e77c8cc2239327c5edb3a432268e5831`
-
-const AAVE_ARB_USDC_TOKEN_ID_STAGING = `${NetworkId['arbitrum-sepolia']}:0x460b97bd498e1157530aeb3086301d5225b91216`
-const AAVE_ARB_USDC_TOKEN_ID_MAINNET = `${NetworkId['arbitrum-one']}:0x724dc807b04555b71ed48a6896b6f41593b8c637`
-
-const ARB_AAVE_POOL_V3_CONTRACT_ADDRESS_STAGING = '0xBfC91D59fdAA134A4ED45f7B584cAf96D7792Eff'
-const ARB_AAVE_POOL_V3_CONTRACT_ADDRESS_MAINNET = '0x794a61358D6845594F94dc1DB02A252b5b4814aD'
-
-// also called rewards controller
-const ARB_AAVE_INCENTIVES_V3_CONTRACT_ADDRESS_STAGING = '0x3A203B14CF8749a1e3b7314c6c49004B77Ee667A'
-const ARB_AAVE_INCENTIVES_V3_CONTRACT_ADDRESS_MAINNET = '0x929EC64c34a17401F460460D4B9390518E5B473e'
 
 const CLOUD_FUNCTIONS_STAGING = 'https://api.alfajores.valora.xyz'
 const CLOUD_FUNCTIONS_MAINNET = 'https://api.mainnet.valora.xyz'
@@ -418,9 +405,6 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     spendTokenIds: [CUSD_TOKEN_ID_STAGING, CEUR_TOKEN_ID_STAGING],
     saveContactsUrl: SAVE_CONTACTS_ALFAJORES,
     getPointsConfigUrl: GET_POINTS_CONFIG_ALFAJORES,
-    arbAavePoolV3ContractAddress: ARB_AAVE_POOL_V3_CONTRACT_ADDRESS_STAGING,
-    arbAaveIncentivesV3ContractAddress: ARB_AAVE_INCENTIVES_V3_CONTRACT_ADDRESS_STAGING,
-    aaveArbUsdcTokenId: AAVE_ARB_USDC_TOKEN_ID_STAGING,
     internalRpcUrl: {
       [Network.Arbitrum]: INTERNAL_ARBITRUM_RPC_URL_STAGING,
     },
@@ -520,9 +504,6 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     spendTokenIds: [CUSD_TOKEN_ID_MAINNET, CEUR_TOKEN_ID_MAINNET],
     saveContactsUrl: SAVE_CONTACTS_MAINNET,
     getPointsConfigUrl: GET_POINTS_CONFIG_MAINNET,
-    arbAavePoolV3ContractAddress: ARB_AAVE_POOL_V3_CONTRACT_ADDRESS_MAINNET,
-    arbAaveIncentivesV3ContractAddress: ARB_AAVE_INCENTIVES_V3_CONTRACT_ADDRESS_MAINNET,
-    aaveArbUsdcTokenId: AAVE_ARB_USDC_TOKEN_ID_MAINNET,
     internalRpcUrl: {
       [Network.Arbitrum]: INTERNAL_ARBITRUM_RPC_URL_MAINNET,
     },
